% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timemoir.R
\name{as_timemoir}
\alias{as_timemoir}
\title{Coerce to a timemoir object}
\usage{
as_timemoir(x)
}
\arguments{
\item{x}{Object to be coerced}
}
\description{
This is typically needed only if you are performing additional manipulations
after calling \code{\link[=timemoir]{timemoir()}}.
}
