% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.timemoir}
\alias{autoplot.timemoir}
\alias{plot.timemoir}
\title{Autoplot method for timemoir results (grouped bar chart)}
\usage{
autoplot.timemoir(object, unit_memory = c("MB", "GB"), ...)

\method{plot}{timemoir}(x, unit_memory = c("MB", "GB"), ..., y)
}
\arguments{
\item{object}{A result from timemoir()}

\item{unit_memory}{unit for memory display: "MB" or "GB" (default to "MB")}

\item{...}{unused}

\item{x}{A \code{timemoir} object.}

\item{y}{Ignored, required for compatibility with the \code{plot()} generic.}
}
\value{
A ggplot object with grouped bars (duration, memory, cpu)
}
\description{
Autoplot method for timemoir results (grouped bar chart)
}
\details{
This function requires some optional dependencies. \link[ggplot2:ggplot2-package]{ggplot2},
\link[tidyr:tidyr-package]{tidyr}, and \link[dplyr:dplyr-package]{dplyr}.
}
\examples{
test_function <- function(n) {
  x <- rnorm(n)
  mean(x)
}

res <- timemoir(test_function(1.2e7), test_function(1.5e7), test_function(1e7))

if (require(ggplot2) && require(tidyr) && require(dplyr)) {
  plot(res)
}
}
