% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timemoir.R
\name{launch_function}
\alias{launch_function}
\title{launch_function}
\usage{
launch_function(xfun, flag_file = tempfile())
}
\arguments{
\item{xfun}{the function to test}

\item{flag_file}{the flag file to use to check xfun has finished}
}
\value{
a named list with `max_mem`, the max used memory, `duration` the
  duration of the function, result, the result of the function and `error` if
  function exits with error
}
\description{
launch xfun in background and watch the pid file to get max memory
}
\examples{
my_fun <- function(sec) {
  Sys.sleep(sec)
  return(TRUE)
}

launch_function(my_fun(1))

my_fun <- function(sec) {
  Sys.sleep(sec)
  return(TRUE)
}

launch_function(my_fun())
}
