% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timemoir.R
\name{timemoir}
\alias{timemoir}
\title{Benchmark functions}
\usage{
timemoir(..., verbose = TRUE, n = 1, interval = 0.1)
}
\arguments{
\item{...}{functions to benchmark.}

\item{verbose}{A boolean. If TRUE (default) print information messages.}

\item{n}{An integer. number of time each function must run (default 1)}

\item{interval}{(default 0.1) sleep interval between memory check in sec}
}
\value{
A result tibble with one row per benchmarked function and 7 columns:
\itemize{
\item \code{fname}, function name (as string).
\item \code{duration} duration (in sec) of the function or NA if function fails.
\item \code{error}, error message if function fails, NA otherwise.
\item \code{start_mem} memory used before function benchmark (in KB).
\item \code{max_mem} max used memory (in KB).
\item \code{cpu_user} user cpu used in sec
\item \code{cpu_sys} system cpu used in sec
}

An error message such as "Process is a zombie" typically indicates that
the process was terminated after exceeding the allocated memory quota.
}
\description{
Executes each function in a separate forked process and monitors its memory and cpu usage in real time

This function is best used with long running functions like \code{arrow} or
\code{duckdb} requests that doesn't fit with classic benchmarking methods like
\code{utils::Rprof} and \code{profmem}.

memory is extracted every \code{interval} sec with \code{ps::ps_memory_info}
\itemize{
\item \code{start_mem} is measured just before launching the function.
\item \code{max_mem} is the max of all measured mem of the forked child
}

cpu times and duration are given by \code{proc.time}
}
\examples{

test_function <- function(n) {
  x <- rnorm(n); mean(x)
}
timemoir(Sys.sleep(2), Sys.sleep(), test_function(1e7))
}
