% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timemoir.R
\name{timemoir}
\alias{timemoir}
\title{timemoir}
\usage{
timemoir(...)
}
\arguments{
\item{...}{functions to benchmark.}
}
\value{
a tibble with `fname`, the function name (as a string) passed to
  `timemoir`, max_mem` the max used memory, `duration` the duration of the
  function (or NA if function fails) and `error`, the error message if
  function fails (or NA if function pass)
}
\description{
launch xfun in background and watch the pid file to get max memory
}
\examples{
timemoir(Sys.sleep(1), Sys.sleep())
}
