% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timemoir.R
\name{timemoir}
\alias{timemoir}
\title{Benchmark functions}
\usage{
timemoir(..., verbose = TRUE)
}
\arguments{
\item{...}{functions to benchmark.}

\item{verbose}{A boolean. If TRUE (default) print information messages.}
}
\value{
a tibble with `fname`, the function name (as a string) passed to
  `timemoir`, max_mem` the max used memory, `duration` the duration of the
  function (or NA if function fails) and `error`, the error message if
  function fails (or NA if function pass)
}
\description{
launch functions in background and watch the pid file to get max used memory
}
\examples{
timemoir(Sys.sleep(2), Sys.sleep())

timemoir(Sys.sleep(1), Sys.sleep(), verbose=FALSE)
}
