% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timemoir.R
\name{timemoir}
\alias{timemoir}
\title{timemoir}
\usage{
timemoir(xfun, flag_file = tempfile())
}
\arguments{
\item{xfun}{the function to test}

\item{flag_file}{the flag file to use to check xfun has finished}
}
\value{
a named list with `fname`, the function name (as a string) passed to
  `timemoir`, max_mem` the max used memory, `duration` the duration of the
  function (or NA if function fails) and `error`, the error message if
  function fails (or NA if function pass)
}
\description{
launch xfun in background and watch the pid file to get max memory
}
\examples{
my_fun <- function(sec) {
  Sys.sleep(sec)
  return(TRUE)
}

timemoir(my_fun(1))

my_fun <- function(sec) {
  Sys.sleep(sec)
  return(TRUE)
}

timemoir(my_fun())
}
